feat: upload widget with a simple example
